fix AddressSection
