--- client/app/(server)/api/locations/route.ts
@@ -1,9 +1,0 @@
-import { NextResponse } from 'next/server'
-
-import { getAllLocations } from './data'
-
-// Notice the function definition:
-export async function GET() {
-  // ...
-  return NextResponse.json(getAllLocations())
-}
